float ptg start

+++ main.py
@@ -31,7 +31,7 @@ def run_and_eval (n_samples : Annotated[Optional[int], typer.Option(help="Number
           max_length: Annotated[Optional[int], typer.Option(help="Maximum length of input sequences")] = 512,
           pruning_schedule : Annotated[Optional[str], typer.Option(help="Pruning schedule - can be agp or linear")] = "linear",
           prune_every_epoch : Annotated[Optional[int], typer.Option(help="Prune at every n epoch")] = 1,
-          start_pruning_epoch_ptg : Annotated[Optional[int], typer.Option(help="Start pruning at n percent epoch of training - will be rounded down")] = 0):
+          start_pruning_epoch_ptg : Annotated[Optional[float], typer.Option(help="Start pruning at n percent epoch of training - will be rounded down")] = 0):
     
     pruning_method = "L1Unstructured"
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -46,7 +46,7 @@ def run_and_eval (n_samples : Annotated[Optional[int], typer.Option(help="Number
         pruning_schedule = "linear"
 
     pruning_start_epoch = int(np.floor(num_epochs * start_pruning_epoch_ptg))
-    
+
     if pruning_start_epoch >= num_epochs:
         pruning_start_epoch = num_epochs - 1
